+++ nuxeo-drive-client/nxdrive/engine/processor.py
@@ -146,6 +146,9 @@ class Processor(EngineWorker):
                 if not self.check_pair_state(doc_pair):
                     continue
 
+                self._engine._manager.osi.send_sync_status(
+                    doc_pair, local_client.abspath(doc_pair.local_path))
+
                 if (AbstractOSIntegration.is_mac()
                         and local_client.exists(doc_pair.local_path)):
                     try:
@@ -224,6 +227,11 @@ class Processor(EngineWorker):
                 try:
                     soft_lock = self._lock_soft_path(doc_pair.local_path)
                     sync_handler(doc_pair, local_client, remote_client)
+
+                    pair = self._dao.get_state_from_id(doc_pair.id)
+                    if pair and 'deleted' not in pair.pair_state:
+                        self._engine._manager.osi.send_sync_status(
+                            pair, local_client.abspath(pair.local_path))
                     self._current_metrics['end_time'] = current_milli_time()
                     self.pairSync.emit(doc_pair, self._current_metrics)
                 except ThreadInterrupt: