+++ nuxeo-drive-client/nxdrive/osi/darwin/pyNotificationCenter.py
@@ -1,29 +1,31 @@
 # coding: utf-8
-""" Python integration with Mountain Lion's notification center. """
-
-import objc
-
-NSUserNotification = objc.lookUpClass('NSUserNotification')
-NSUserNotificationCenter = objc.lookUpClass('NSUserNotificationCenter')
-NSObject = objc.lookUpClass('NSObject')
+""" Python integration macOS notification center. """
+from Foundation import (NSBundle, NSMutableDictionary, NSObject,
+                        NSUserNotification, NSUserNotificationCenter)
+from ...constants import BUNDLE_IDENTIFIER
 
 
 class NotificationDelegator(NSObject):
 
     def __init__(self):
         self._manager = None
+        info_dict = NSBundle.mainBundle().infoDictionary()
+        if not 'CFBundleIdentifier' in info_dict:
+            info_dict['CFBundleIdentifier'] = BUNDLE_IDENTIFIER
 
-    def userNotificationCenter_didActivateNotification_(self, center, notification):
+    def userNotificationCenter_didActivateNotification_(self, center,
+                                                        notification):
         info = notification.userInfo()
-        if "uuid" not in info or self._manager is None:
+        if 'uuid' not in info or self._manager is None:
             return
         notifications = self._manager.notification_service.get_notifications()
-        if info['uuid'] not in notifications or (info['uuid'] in notifications and
-                                                 notifications[info['uuid']].is_discard_on_trigger()):
+        if (info['uuid'] not in notifications
+                or notifications[info['uuid']].is_discard_on_trigger()):
             center.removeDeliveredNotification_(notification)
-        self._manager.notification_service.trigger_notification(info["uuid"])
+        self._manager.notification_service.trigger_notification(info['uuid'])
 
-    def userNotificationCenter_shouldPresentNotification_(self, center, notification):
+    def userNotificationCenter_shouldPresentNotification_(self, center,
+                                                          notification):
         return True
 
 
@@ -43,12 +45,12 @@ def notify(title, subtitle, info_text, delay=0, sound=False, user_info=None):
         userInfo: a dictionary that can be used to handle clicks in your
                   app's applicationDidFinishLaunching:aNotification method
     """
-    if user_info is None:
-        user_info = {}
     notification = NSUserNotification.alloc().init()
     notification.setTitle_(title)
     notification.setSubtitle_(subtitle)
     notification.setInformativeText_(info_text)
+    user_info = user_info or {}
+    user_info = NSMutableDictionary.alloc().init().setDictionary_(user_info)
     notification.setUserInfo_(user_info)
     if sound:
         notification.setSoundName_('NSUserNotificationDefaultSoundName')